Rename orkestrator seksjon og legg tilbake tracking av fullført seksjon
