+++ src/views/soknad/Soknad.tsx
@@ -9,7 +9,7 @@ import { PageMeta } from "../../components/PageMeta";
 import { Personalia } from "../../components/personalia/Personalia";
 import { ProgressBar } from "../../components/progress-bar/ProgressBar";
 import { SoknadHeader } from "../../components/soknad-header/SoknadHeader";
-import { OrkestratorSection } from "../../components/section/Section";
+import { OrkestratorSection } from "../../components/section/OrkestratorSection";
 import { useSoknad } from "../../context/soknad-context";
 import { useSanity } from "../../context/sanity-context";
 import { useValidation } from "../../context/validation-context";
@@ -19,7 +19,7 @@ import styles from "./Soknad.module.css";
 import { ErrorTypesEnum } from "../../types/error.types";
 import { QuizSection } from "../../components/section/QuizSection";
 import { getUnansweredFaktumId } from "../../components/faktum/validation/validations.utils";
-// import { trackSkjemaStegFullført } from "../../amplitude.tracking";
+import { trackSkjemaStegFullført } from "../../amplitude.tracking";
 
 interface IProps {
   personalia: IPersonalia | null;
@@ -29,7 +29,7 @@ export function Soknad(props: IProps) {
   const router = useRouter();
   const { getAppText } = useSanity();
   const { totalSteps } = useProgressBarSteps();
-  const { quizState: soknadState, orkestratorState, isError, isLoading, isLocked } = useSoknad();
+  const { quizState, orkestratorState, isError, isLoading, isLocked } = useSoknad();
   const { unansweredFaktumId, setUnansweredFaktumId } = useValidation();
   const [navigating, setNavigating] = useState(false);
 
@@ -38,7 +38,6 @@ export function Soknad(props: IProps) {
   const invalidSectionParams = isNaN(Number(sectionParams));
   const sectionNumber = invalidSectionParams ? 1 : Number(sectionParams);
   const sectionIndex = sectionNumber - 1;
-  const isLastSection = soknadState.ferdig && orkestratorState && orkestratorState.erFullført;
 
   // Number of sections
   const numberOfOrkestratorSections = (orkestratorState && orkestratorState.antallSeksjoner) || 0;
@@ -47,21 +46,21 @@ export function Soknad(props: IProps) {
   // Orkestrator and Quiz section data based on sectionParams
   const currentOrkestratorSectionData =
     orkestratorState && orkestratorState.seksjoner[sectionIndex];
-  const currentQuizSectionData = soknadState.seksjoner[sectionIndex - numberOfOrkestratorSections];
+  const currentQuizSectionData = quizState.seksjoner[sectionIndex - numberOfOrkestratorSections];
 
   // Validation
-  const firstUnansweredSectionIndex = soknadState.seksjoner.findIndex((seksjon) => !seksjon.ferdig);
+  const firstUnansweredSectionIndex = quizState.seksjoner.findIndex((seksjon) => !seksjon.ferdig);
   const firstUnfinishedSection = firstUnansweredSectionIndex + 1;
 
   // Teste commit
   useEffect(() => {
     // Automatisk redirect til siste ubesvart seksjon dersom man kommer fra inngang siden
-    if (router.query.fortsett && !soknadState.ferdig && firstUnansweredSectionIndex !== -1) {
+    if (router.query.fortsett && !quizState.ferdig && firstUnansweredSectionIndex !== -1) {
       router.push(`/soknad/${router.query.uuid}?seksjon=${firstUnfinishedSection}`);
     }
 
     const availiableSections =
-      soknadState.seksjoner.length + ((orkestratorState && orkestratorState?.antallSeksjoner) || 0);
+      quizState.seksjoner.length + ((orkestratorState && orkestratorState?.antallSeksjoner) || 0);
 
     if (invalidSectionParams || sectionNumber > availiableSections) {
       // Hvis vi ikke finner en seksjon så sender vi bruker automatisk til første seksjon
@@ -75,11 +74,12 @@ export function Soknad(props: IProps) {
     if (unansweredFaktumId) {
       setUnansweredFaktumId(undefined);
     }
-  }, [soknadState, orkestratorState]);
+  }, [quizState, orkestratorState]);
 
   function navigateToNextSection() {
     if (isOrkestratorSection) {
       if (currentOrkestratorSectionData && !currentOrkestratorSectionData.erFullført) {
+        trackSkjemaStegFullført("dagpenger", router.query.uuid as string, sectionNumber);
         setUnansweredFaktumId(currentOrkestratorSectionData.nesteUbesvarteOpplysning.opplysningId);
       } else {
         const nextIndex = sectionNumber + 1;
@@ -91,6 +91,7 @@ export function Soknad(props: IProps) {
 
     if (!isOrkestratorSection) {
       if (!currentQuizSectionData.ferdig) {
+        trackSkjemaStegFullført("dagpenger", router.query.uuid as string, sectionNumber);
         const unansweredFaktumId = getUnansweredFaktumId(currentQuizSectionData.fakta);
         setUnansweredFaktumId(unansweredFaktumId);
       } else {
@@ -165,8 +166,7 @@ export function Soknad(props: IProps) {
             </Button>
           )}
 
-          {/* Denne her er kanskje ikke helt riktig, fiks det  */}
-          {sectionNumber === totalSteps - 2 && isLastSection && soknadState.ferdig ? (
+          {orkestratorState.erFullført && quizState.ferdig ? (
             <Button onClick={() => navigateToDocumentation()} loading={navigating}>
               {getAppText("soknad.knapp.til-dokumentasjon")}
             </Button>
